+++ src/hcsr04/hcsr04.cxx
@@ -38,7 +38,7 @@
 
 using namespace upm;
 
-HCSR04::HCSR04 (uint8_t triggerPin, uint8_t echoPin) {
+HCSR04::HCSR04 (int triggerPin, int echoPin) {
     m_name              = "HCSR04";
 
     m_triggerPinCtx     = mraa_gpio_init (triggerPin);
